# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for toy rental service",
        "title": "ToyRental API",
        "contact": {
            "name": "Indrawansyah",
            "email": "indra@dev.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/business-report/customers": {
            "get": {
                "description": "Mendapatkan laporan pelanggan yang paling aktif berdasarkan jumlah penyewaan dan pengeluaran",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business Report"
                ],
                "summary": "Mendapatkan laporan pelanggan teratas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tanggal mulai (YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tanggal akhir (YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Jumlah data (default: 10)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.APISuccessResponse"
                        }
                    }
                }
            }
        },
        "/business-report/popular-toys": {
            "get": {
                "description": "Mendapatkan laporan mainan paling populer berdasarkan jumlah penyewaan",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business Report"
                ],
                "summary": "Mendapatkan laporan mainan populer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tanggal mulai (YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tanggal akhir (YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Jumlah data (default: 10)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.APISuccessResponse"
                        }
                    }
                }
            }
        },
        "/business-report/rental-status": {
            "get": {
                "description": "Mendapatkan laporan jumlah penyewaan berdasarkan status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business Report"
                ],
                "summary": "Mendapatkan laporan status penyewaan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tanggal mulai (YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tanggal akhir (YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.APISuccessResponse"
                        }
                    }
                }
            }
        },
        "/business-report/sales": {
            "get": {
                "description": "Mendapatkan laporan penjualan dalam rentang waktu tertentu",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business Report"
                ],
                "summary": "Mendapatkan laporan penjualan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tanggal mulai (YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tanggal akhir (YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pengelompokan (day, week, month)",
                        "name": "group_by",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.APISuccessResponse"
                        }
                    }
                }
            }
        },
        "/payment": {
            "post": {
                "description": "Buat pembayaran baru menggunakan midtrans",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Membuat pembayaran rental",
                "parameters": [
                    {
                        "description": "ID Rental",
                        "name": "rental_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.CreatePaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Payment"
                        }
                    }
                }
            }
        },
...
        "/payment/rental/{rental_id}": {
            "get": {
                "description": "Mendapatkan semua pembayaran berdasarkan ID rental",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Mendapatkan semua pembayaran untuk rental",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Rental",
                        "name": "rental_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Payment"
                            }
                        }
                    }
                }
            }
        },
        "/payment/{id}": {
            "get": {
                "description": "Mendapatkan detail pembayaran berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Mendapatkan detail pembayaran",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Pembayaran",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Payment"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entity.CreatePaymentRequest": {
            "type": "object",
            "required": [
                "rental_id"
            ],
            "properties": {
                "rental_id": {
                    "type": "string"
                }
            }
        },
...
        "entity.Payment": {
            "type": "object",
            "properties": {
                "expiry_time": {
                    "type": "string"
                },
                "fraud_status": {
                    "type": "string"
                },
                "gross_amount": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                },
                "payment_method": {
                    "type": "string"
                },
                "payment_type": {
                    "type": "string"
                },
                "rental_id": {
                    "type": "string"
                },
                "snap_token": {
                    "type": "string"
                },
                "snap_url": {
                    "type": "string"
                },
                "transaction_status": {
                    "type": "string"
                },
                "transaction_time": {
                    "type": "string"
                },
                "va_number": {
                    "type": "string"
                }
            }
        },
...
        "response.APISuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "metadata": {},
                "status_code": {
                    "type": "integer"
                }
            }
        }
    }
}